S03P31A205-76/featureback/board: fix delete update

+++ django_server/photory/board/views.py
@@ -30,25 +30,22 @@ def article_create(request):
         serializer.save()
         return Response(serializer.data)
 
-@api_view(['PUT'])
-@permission_classes([IsAuthenticated])
-def article_update(request,article_pk):
-    article =Article.objects.get(pk=article_pk)
-    article.bookcover = request.data['bookcover']
-    article.title = request.data['title']
-    article.writer = request.data['writer']
-    article.content = request.data['content']
-    article.category = request.data['category']
-    article.save()
-    return Response(status =200)
-
-
-@api_view(['DELETE'])
+@api_view(['DELETE','PUT'])
 @permission_classes([IsAuthenticated]) 
-def article_delete(request, article_pk):
-	article = Article.objects.get(pk = article_pk)
-	article.delete()
-	return Response(status=200)
+def article_delete_update(request, article_pk): 
+    if request.method == 'DELETE':
+        article = Article.objects.get(pk = article_pk)
+        article.delete()
+        return Response(status=200)
+    else:
+        article =Article.objects.get(pk=article_pk)
+        article.bookcover = request.data['bookcover']
+        article.title = request.data['title']
+        article.writer = request.data['writer']
+        article.content = request.data['content']
+        article.category = request.data['category']
+        article.save()
+        return Response(status =200)
 
 def index(request, pagenum,pagenum2):
     articles = Article.objects.order_by('-pk')
